--- a/lib/src/multires_stochastic_texture_synthesis.rs
+++ b/lib/src/multires_stochastic_texture_synthesis.rs
@@ -759,8 +759,9 @@ impl Generator {
                 has_fanned_out = true;
                 let tile_adjusted_width = (self.output_size.width as f32 * 1.1) as u32 + 1;
                 let tile_adjusted_height = (self.output_size.height as f32 * 1.1) as u32 + 1;
-                // heuristic: pick a cell size so that the expected number of resolved points in any cell is k
-                let grid_cell_size = ((params.nearest_neighbors * self.output_size.height * self.output_size.height / redo_count as u32) as f64).sqrt() as u32 + 1;
+                // heuristic: pick a cell size so that the expected number of resolved points in any cell is 4 * k
+                // this seems to be a safe overestimate
+                let grid_cell_size = ((params.nearest_neighbors * self.output_size.height * self.output_size.height / redo_count as u32) as f64).sqrt() as u32 * 2 + 1;
                 let new_tree_grid = TreeGrid::new(
                     tile_adjusted_width,
                     tile_adjusted_height,