@Mr4k 2019-12-04
Sped up the read from the grid of rtrees which brought the runtime down a little:
From a test Google Cloud Machine with 64 threads
```
Command: time cargo run --release -- --out-size 1024  --out out/01.jpg generate imgs/1.jpg
Before grid read speed up:
~7.5 seconds
After grid read speed up:
~5.6 seconds

Command: time cargo run --release -- --out-size 2048 --out out/01.jpg generate imgs/1.jpg
Before grid read speed up:
~28 seconds
After grid read speed up:
~24 seconds
```
But most importantly it now runs a lot faster cpus with few threads.
My development macbook (4 logical cores):
```
Command: time cargo run --release -- --out out/01.jpg generate imgs/1.jpg
Before: ~20s
After: ~11.5s
```