@Jake-Shadle 2019-12-05
Oh and FYI, if you are doing your iterative benchmarking via the cli you might want to check out hyperfine.