@Mr4k 2019-12-17
Fixed in painting. This pr is not mergable yet because it does not work for rectangular images but I will add that likely tomorrow