@Mr4k 2019-12-18
Functionality wise this pr is now complete.
Also took a quick performance benchmark using hyperfine:
Google Cloud 64 core cpu 
```
Before: 
Benchmark #1: ./target/release/texture-synthesis --out-size 2048 --out out/01.jpg generate imgs/1.jpg
  Time (mean ± σ):     64.681 s ±  0.256 s    [User: 734.705 s, System: 104.357 s]
  Range (min … max):   64.293 s … 65.154 s    10 runs

After:
Benchmark #1: ./target/release/texture-synthesis --out-size 2048 --out out/01.jpg generate imgs/1.jpg
  Time (mean ± σ):     20.544 s ±  0.182 s    [User: 796.544 s, System: 9.745 s]
  Range (min … max):   20.362 s … 20.999 s    10 runs
```