@Mr4k 2020-01-08
Thanks so much for reviewing! I'll make the changes soon and generally try to make the naming better! And of course if you get a chance I'd love to know runtimes on more reference rigs but no pressure if you're busy.