@h3r2tic 2020-01-10
Tested on my Threadripper! Pretty consistent with the "Google Cloud 64 core cpu".

```
BEFORE:

Benchmark #1: target\release\texture-synthesis --out-size 2048 --out out/01.jpg generate imgs/1.jpg
  Time (mean ± σ):     62.349 s ±  0.255 s    [User: 1.4 ms, System: 4.4 ms]
  Range (min … max):   61.825 s … 62.664 s    10 runs

AFTER:

Benchmark #1: target\release\texture-synthesis --out-size 2048 --out out/01.jpg generate imgs/1.jpg
  Time (mean ± σ):     23.737 s ±  0.488 s    [User: 3.0 ms, System: 3.6 ms]
  Range (min … max):   23.034 s … 24.373 s    10 runs
```