@Mr4k 2020-01-10
Good to hear! (Although I did accidentally just add and remove a print statement so I think I'll need another approval)

Also just for completeness's sake here are some things I thought of that I decided were out of scope of this pr but might be worth mentioning:
- I'm not sure how common this use case is but for inpainting a few small sparse regions on a large image an sparse version of the tree grid may provide better performance
- In my benchmarking I have noticed that since the contention has been reduced the single threaded first few stages for multithreaded generation (especially with many threads) has become a large bottleneck (up to 50% of the time in some case iirc). A potentially future improvement may be see if that can be some what parallelized especially for larger images
- Also after benchmarking I was also left wondering if an rstar tree is the ideal data structure for this kind of problem (now takes up about 50% of the generation step for a pixel iirc) given that most of the time your k nearest neighbors will be very close to you in predictable positions. I'm not sure what a better data structure looks like but it was just a thing I thought was interesting

Also finally thanks so much @h3r2tic @anopara @Jake-Shadle for taking the time to benchmark / review my code! It was really fun to work on this.